For Horizon, making it so the cookies during authentication are preserved for the Horizon app so that they can later be reused to reauthenticate the user when trying to view inbox and skillspace. It's not ready for merge yet however because we need to destroy those cookies when the user logs out [ignore-commit-lint]

+++ Core/Core/Common/CommonModels/AppEnvironment/AppEnvironment.swift
@@ -91,13 +91,15 @@ open class AppEnvironment {
         refreshWidgets()
         saveAccount(for: session)
 
-        Just(())
-            .receive(on: RunLoop.main)
-            .flatMap { CoreWebView.deleteAllCookies() }
-            .sink {
-                CoreWebView.refreshKeepAliveCookies()
-            }
-            .store(in: &subscriptions)
+        if AppEnvironment.shared.app != .horizon {
+            Just(())
+                .receive(on: RunLoop.main)
+                .flatMap { CoreWebView.deleteAllCookies() }
+                .sink {
+                    CoreWebView.refreshKeepAliveCookies()
+                }
+                .store(in: &subscriptions)
+        }
     }
 
     public func userDidLogout(session: LoginSession) {
